image_api/requirements.txt
fastapi==0.115.0
diffusers==0.30.3
torch==2.4.1
uvicorn==0.30.6
accelerate==0.34.2
pymongo==4.5.0
python-dotenv==1.0.0
